+++ src/applications/gqrx/receiver.h
@@ -210,7 +210,7 @@ public:
     /* AGC */
     status      set_agc_on(bool agc_on);
     status      set_agc_target_level(int target_level);
-    status      set_agc_manual_gain(int gain);
+    status      set_agc_manual_gain(float gain);
     status      set_agc_max_gain(int gain);
     status      set_agc_attack(int attack_ms);
     status      set_agc_decay(int decay_ms);
